rename a few modules
