+++ README.md
@@ -104,11 +104,11 @@ Finally, to simplify computation (and to prevent players from getting lost), Saf
 
 ### Classes and code
 
-All of these rule are encapsulated by the `safelife.game_physics.SafeLifeGame` class. That class is responsible for maintaining the game state associated with each SafeLife level, changing the state in response to player actions, and updating the state at each time step. It also has functions for serializing and de-serializing the state (saving and loading).
+All of these rule are encapsulated by the `safelife.safelife_game.SafeLifeGame` class. That class is responsible for maintaining the game state associated with each SafeLife level, changing the state in response to player actions, and updating the state at each time step. It also has functions for serializing and de-serializing the state (saving and loading).
 
 Actions in `SafeLifeGame` do not typically result in any direct rewards (there is a small bonus for successfully reaching a level exit). Instead, each board state is worth a certain number of points, and agent actions can increase or reduce that point value.
 
-The `safelife.gym_env.SafeLifeEnv` class wraps `SafeLifeGame` in an interface suitable for reinforcement learning agents (à la [OpenAI Gym](https://gym.openai.com/)). It implements `step()` and `reset()` functions. The former accepts an action (integers 0–8) and outputs an observation, reward, whether or not the episode completed, and a dictionary of extra information (see the code for more details); the latter starts a new episode and returns a new observation. Observations in `SafeLifeEnv` are not the same as board states in `SafeLifeGame`. Crucially, the observation is always centered on the agent (this respects the symmetry of the game and means that agents don't have to implement attention mechanisms), can be partial (the agent can only see a certain distance), and only displays the color of the goal cells rather than their full content.
+The `safelife.safelife_env.SafeLifeEnv` class wraps `SafeLifeGame` in an interface suitable for reinforcement learning agents (à la [OpenAI Gym](https://gym.openai.com/)). It implements `step()` and `reset()` functions. The former accepts an action (integers 0–8) and outputs an observation, reward, whether or not the episode completed, and a dictionary of extra information (see the code for more details); the latter starts a new episode and returns a new observation. Observations in `SafeLifeEnv` are not the same as board states in `SafeLifeGame`. Crucially, the observation is always centered on the agent (this respects the symmetry of the game and means that agents don't have to implement attention mechanisms), can be partial (the agent can only see a certain distance), and only displays the color of the goal cells rather than their full content.
 
 The reward function in `SafeLifeEnv` is just the (rescaled) difference in point values between the board before and after an action and time-step update. In addition, `SafeLifeEnv` introduces a no-movement penalty to encourage agent action, and an unobserved time limit to prevent agents from getting stuck. When the environment is reset, either a new board is chosen from a set of fixed levels, or a new level is randomly generated.
