+++ requirements.txt
@@ -5,3 +5,4 @@ gym>=0.12.5
 imageio>=2.5.0
 ipython>=5.0
 pyglet==1.3.2
+pyyaml>=3.12
