+++ safelife/benchmarking.py
@@ -4,14 +4,14 @@ from types import SimpleNamespace
 import yaml
 import numpy as np
 
-from safelife.file_finder import safelife_loader
-from safelife.gym_env import SafeLifeEnv
-from . import wrappers
+from .file_finder import safelife_loader
+from .safelife_env import SafeLifeEnv
+from . import env_wrappers
 
 
 def run_benchmark(
         name, policy, logfile, num_trials=1, record=True, num_env=10,
-        env_args={}):
+        env_factory=SafeLifeEnv):
     """
     Run benchmark levels for a specific policy.
 
@@ -35,6 +35,10 @@ def run_benchmark(
         as the log file.
     num_env : int, optional
         Number of environments to run simultaneously.
+    env_factory : function
+        Function to build new SafeLifeEnv instances. Must accept at least two
+        arguments: ``level_iterator`` and ``global_counter``. See
+        :class:`safelife_env.SafeLifeEnv` for more details.
     """
     logfile = os.path.abspath(os.path.expanduser(logfile))
     logdir = os.path.split(logfile)[0]
@@ -60,11 +64,11 @@ def run_benchmark(
 
     envs = []
     obs = []
-    rnn_state = None
+    rnn_state = [None] * num_env
     for k in range(num_env):
-        env = SafeLifeEnv(levels, global_counter=counter, **env_args)
+        env = env_factory(level_iterator=levels, global_counter=counter)
         # Note that basically all the logging happens in the wrapper.
-        env = wrappers.RecordingSafeLifeWrapper(
+        env = env_wrappers.RecordingSafeLifeWrapper(
             env, log_file=logfile, video_name=video_name, video_recording_freq=1)
         try:
             obs.append(env.reset())
@@ -90,6 +94,7 @@ def run_benchmark(
                 if done:
                     try:
                         ob = env.reset()
+                        state = None
                     except StopIteration:
                         continue
                 new_obs.append(ob)
