+++ safelife/safelife_game.py
